LPS-125095 Move classes that implement FormNavigatorCategory to package *.frontend.taglib.form.navigator
