--- modules/apps/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/internal/servlet/taglib/ui/SubscriptionsFormNavigatorCategory.java
@@ -1,50 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.asset.publisher.web.internal.servlet.taglib.ui;
-
-import com.liferay.asset.publisher.constants.AssetPublisherConstants;
-import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
-import com.liferay.portal.kernel.language.LanguageUtil;
-
-import java.util.Locale;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Eudaldo Alonso
- */
-@Component(
-	property = "form.navigator.category.order:Integer=10",
-	service = FormNavigatorCategory.class
-)
-public class SubscriptionsFormNavigatorCategory
-	implements FormNavigatorCategory {
-
-	@Override
-	public String getFormNavigatorId() {
-		return AssetPublisherConstants.FORM_NAVIGATOR_ID_CONFIGURATION;
-	}
-
-	@Override
-	public String getKey() {
-		return AssetPublisherConstants.CATEGORY_KEY_SUBSCRIPTIONS;
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "subscriptions");
-	}
-
-}
