--- modules/apps/commerce/commerce-product-type-virtual-web/src/main/java/com/liferay/commerce/product/type/virtual/web/internal/servlet/taglib/ui/CPDefinitionVirtualSettingDetailsFormNavigatorCategory.java
@@ -1,52 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.commerce.product.type.virtual.web.internal.servlet.taglib.ui;
-
-import com.liferay.commerce.product.type.virtual.web.internal.servlet.taglib.ui.constants.CPDefinitionVirtualSettingFormNavigatorConstants;
-import com.liferay.frontend.taglib.form.navigator.FormNavigatorCategory;
-import com.liferay.portal.kernel.language.LanguageUtil;
-
-import java.util.Locale;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Alessio Antonio Rendina
- */
-@Component(
-	enabled = false, property = "form.navigator.category.order:Integer=100",
-	service = FormNavigatorCategory.class
-)
-public class CPDefinitionVirtualSettingDetailsFormNavigatorCategory
-	implements FormNavigatorCategory {
-
-	@Override
-	public String getFormNavigatorId() {
-		return CPDefinitionVirtualSettingFormNavigatorConstants.
-			FORM_NAVIGATOR_ID_CP_DEFINITION_VIRTUAL_SETTING;
-	}
-
-	@Override
-	public String getKey() {
-		return CPDefinitionVirtualSettingFormNavigatorConstants.
-			CATEGORY_KEY_CP_DEFINITION_VIRTUAL_SETTING_DETAILS;
-	}
-
-	@Override
-	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "details");
-	}
-
-}
